Split up Comment into individual rules
We can't describe NestingBlockComment using tree-sitter grammar alone,
so it must be an external.

Because we can't reference other definitions inside token definitions,
the choice must be hoisted out of the token(...) rule.

Unfortunately simply defining $.comment as a choice between the
various comment types doesn't work due to a tree-parser bug, so these
are "inlined" in the extras array directly.