[generator] Introduce and use SeqChoice, an intermediate representation node
SeqChoice is a node used in an intermediate internal representation of
the grammar tree, and is a superset of Choice, Seq and Optional.
Using a single node for this purpose allows writing more generic
transformation rules and optimizations.

The grammar.js changes reflect the effect of the improved
optimizations.  Though the optimization and refactoring algorithms
themselves haven't changed in their goals, the new representation
allows them to be much more effective.