[generator] grammar: Lift prefixes and suffixes from choice subsets, too
This is a generalization of the old logic which dealt with the
zero-length ("optional") choice.

If we find two choices which have the same (non-empty) prefix and
suffix, we lift it.

This greatly reduces the size of the generated grammar (mostly in the
inlined token definitions), but also enables more better
transformations (e.g. ParameterAttributes is now in a form which will
be much easier to de-recurse).