Mostly fix float literals + UFCS
This is pretty ugly, due to tree-sitter not being able to either do
look-ahead or disambiguate different ways to lex characters into
tokens (e.g. "5." is either a FloatLiteral or an IntegerLiteral as
part of a PostfixExpression).

It is still not fully correct, but doing so would require moving the
entirety of lexing IntegerLiteral and FloatLiteral into the C++
scanner, to take advantage of its lookahead.