test: Add dmd/runnable tests
Though we succeeded in parsing all tests in DMD's "compilable" corpus,
there are still a few tests in its "runnable" corpus that we cannot
parse.

Of course, all tests which are "runnable" should also be "compilable",
so we should pay attention to these too.

We should take care of these before we move on to attempting to fix
issues in community projects, as doing things in this order will allow
us to more accurately assess if a syntactic construct seen in the wild
really wasn't covered by DMD's test suite (and therefore is more
likely to indicate a hole in DMD's test coverage).